tests/wpt/web-platform-tests/webmessaging/Channel_postMessage_clone_port_error.htm

<!DOCTYPE html>
<html>
<head>
<title> postMessage() DataCloneError: cloning source port </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>

    var description = "Test Description: Throw a DataCloneError if transfer array in postMessage contains source port.";

    test(function()
    {
        var channel = new MessageChannel();
        channel.port1.start();

        assert_throws_dom("DATA_CLONE_ERR", function()
        {
            channel.port1.postMessage("ports", [channel.port1]);
        });
    }, description);
</script>
</body>
</html>
